+++ tools/sdk/lwip2/builder/Makefile.lwip2
@@ -10,6 +10,7 @@ SRC = \
 	$(patsubst %.c,$(BUILD)/%.o,$(wildcard core/ipv4/*.c)) \
 	$(patsubst %.c,$(BUILD)/%.o,$(wildcard core/ipv6/*.c)) \
 	$(patsubst %.c,$(BUILD)/%.o,$(wildcard api/*.c)) \
+	$(patsubst %.c,$(BUILD)/%.o,$(wildcard apps/sntp/*.c)) \
 
 BUILD_INCLUDES = -I$(BUILD) -I$(SDK)/include -Iinclude -I../../glue -I../../glue-lwip2
 
@@ -19,5 +20,6 @@ include $(ROOT)/Makefile.common
 -include $(BUILD)/core/ipv4/*.d
 -include $(BUILD)/core/ipv6/*.d
 -include $(BUILD)/api/*.d
+-include $(BUILD)/apps/sntp/*.d
 
 jall: $(LWIP_LIB)
