+++ tools/sdk/lwip2/include/README.md
@@ -1,0 +1,1 @@
+warning: this directory is re/over/written from ../builder/ files upon lwip2 build
