+++ tools/sdk/lwip2/include/lwipopts.h
@@ -1,3 +1,4 @@
+/* this will be overwritten from ../builder/glue-lwip2/lwip2/lwipopts.h */
 
 #ifndef MYLWIPOPTS_H
 #define MYLWIPOPTS_H
@@ -2967,6 +2968,20 @@
 /**
  * @}
  */
+
+
+/*
+   --------------------------------------------------
+   ------------------ SNTP options ------------------
+   --------------------------------------------------
+*/
+#define SNTP_SERVER_DNS            1	// SNTP support DNS names through sntp_setservername / sntp_getservername
+
+
+/*
+   --------------------------------------------------
+   --------------------------------------------------
+*/
 
 #include "lwip/debug.h"
 #include "arch/cc.h"
